+++ apple_store/Device.py
@@ -1,0 +1,15 @@
+#! /usr/bin/python3
+# -*- coding: utf-8 -*-
+
+class OriginalRefurbishedDevice:
+    ''' Реалізація оригінального відновленого приладу '''
+    
+    def __str__(self) -> str:
+        return "Оригінальний відновлений пристрій"
+
+
+class ReplicaDevice:
+    ''' Реалізація репліки з одеського підвалу '''
+    
+    def __str__(self) -> str:
+        return "Репліка з одеського підвалу"
